backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
anthropic==0.34.0
supabase==2.7.4
voyageai==0.2.3
python-dotenv==1.0.1
wheel==0.46.3
